developer-build.sh: tarballs and PDFs in separate server directories
Until now, the `developer-build.sh' script would rename the tarball to
`gnuastro-latest.tar.lz' and upload the tarball and PDF file to the top
directory of the server. But for archival reasons, it is commonly necessary
to keep the tarball versions on the server. So with this commit, the
uploaded tarball will keep its version number and will be put in the `src'
directory of the server and the script will run SSH after copying the files
and make a symbolic link to the uploaded tarball that will always point to
the most recent version.