Zlib check added before libcurl for static builds
When building statically, libcurl depends on linking with Zlib. To allow
for such senarios, we thus need to add the `-lz' flag to the list of `LIBS'
before we check for libcurl. This won't cause a problem in other scenarios,
and Zlib isn't an optional or mandatory dependency of Gnuastro.