Catalog read before finalizing output size in MakeProfiles
Due to some previous changes in the way MakeProfiles reads columns and
prepares the output array, we were mistakenly finalizing the output size
before actually reading the column contents. Therefore the `--prepforconv'
option (which needs to see if a PSF profile was requested and also needs
its size) had effectively become useless.

Since reading of the columns needs knowledge of the output's
dimensionality, the steps to find the dimensionality are now separated into
a separate function that is done before reading columns. After the columns
are read, we'll go onto finalizing the output's array size.

Fernando Buitrago reported this bug.

This fixes bug #55025.