Atleast one width is required for PDF or EPS outputs in ConvertType
When none of the width options (`--borderwidth', or `--widthincm') are set
and an EPS or PDF output is desired, until now, ConvertType wouldn't
complain and produces a blank page (the size of the output becomes zero, or
nothing)! This could cause a lot of confusion! So a check has been added
when these formats are requested informing the user to atleast provide one
of these width values.