Sanity check added on value to --section in Crop
Until now, there was no sanity checks on the values given to the
`--section' option in Crop! Therefore if (for example) there are alphabetic
characters in the value, it will fall into an infinite loop and hang! With
this commit, if any unrecognized character is encountered, Crop will abort
with an informative description.

This fixes bug #55157.