Annum as unit of year in abbreviated CosmicCalculator units
Until now, we were simply using the non-standard "Gyr" for representing
billion years. However, the official IAU-recognized abbreviation for year
is "a" (for annum), and it is widely used in other scientific branches like
geology, palaeontology, etc. The IAU published units are available here:
https://www.iau.org/publications/proceedings_rules/units/

Therefore with this commit, all mentions of "Gyr" in CosmicCalculator were
corrected to "Ga". A short explanation is also added for those that may not
be familiar with this standard: In the output of `--help', it is written as
"Ga: Giga Annum". In the output with all measurements it is written as
(Ga*), with the star pointing to a footnote in the end. And finally, in the
book its written as "Ga (Giga annum, or billion years)".

This issue was raised up by David Valls-Gabaud.