gal_convolve_spatial ignores convoverch when input isn't a tile
When the input isn't a tile, the `convoverch' argument to
`gal_convolve_spatial' must be 1. Otherwise it will crash (and really
confuse a user who is giving a simple complete dataset instead of a
tessellation).

With this commit, in such scenarios we manually set to the value of
`convoverch' to 1 and ignore the user's given value (most probably zero in
that scenario).