No extra space after Statistics single valued measurements
Until now, when printing the sigle-valued measurements of Statistics, we
would simply print it with an empty space afterwards. This would allow
easily printing any number of measurements after each other.

However, in some situations (in scripts), this white space can cause
problems when storing the output. So with this commit, a counter is added
to the printing function and a white space is printed before each new
measurement if it isn't the first one.