New gal_convolve_spatial_general sanity check
When the input dataset is part of a linked list, but they don't point to
any block, `gal_convolve_spatial_general' will crash because it is
optimized to deal with a list of tiles. So with this commit, a sanity check
is added to check for this scenario and inform the user and abort.