Arithmetic only collapses WCS once
Until now, when there was more than one collapse operator, Arithmetic will
crash with an error in `gal_wcs_remove_dimension' complaining about having
a 2-dimensional WCS. This was because in Arithmetic, we only have one WCS
structure and we were collapsing it two times.

To fix the issue, an integer is added in the main structure to mark if the
WCS structure has already been collapsed or not. The WCS collapse will only
happen if it hasn't already been done.