Maximum unsigned type numbers as string read into larger type
Until now, when you wanted to select pixels with a label of 255 (for
example on NoiseChisel's output) with Arithmetic, it doesn't select
them. This is because 255 fits into an unsigned 8-bit value, so it is
stored with that type. But being the maximum value, 255 is also the value
used for blank in this type.

With this commit, when the user gives 255 to Arithmetic (and in particular
the library's `gal_type_string_to_number' function), it will be put into
the larger type to avoid such situations. The same fix is implemented for
the 16-bit and 32-bit unsigned types.