Preparations for the 8th Gnuastro release
All the necessary preparations have been made for the 8th release of
Gnuastro.

In particular some typos were found: in a convolve library's error message
a typo was found by Lintian when checking for the new release of Gnuastro
in Debian's package manager. Also I went over the new parts of the book and
fixed some typos.

Also, some corrections were made in the release checklist.