No extra blank keyword line after writing WCS keywords
After writing all the WCS keywords, WCSLIB writes a blank keyword record
and a `COMMENT' to mention its version. But Gnuastro writes the version of
WCSLIB afterwards, so to keep things clean we would remove the last line.

However, WCSLIB also writes a blank line before the `COMMENT' which would
cause unequal spacing between the groups of keywords. So with this commit,
instead of just ignoring the last line, we won't write those that are blank
or those that have a `COMMENT' keyword.