Fits's --write now has title and checksum special keywords
Until now, the `--write' option of the FITS program would only write the
given values to a keyword in the FITS header. But with this commit, we have
added the infra-structure to include special keywords: for example those
witout a value (a title), or options where the value must be calculated
internally (checksum).

This also fixes bug #55333.