Minor corrections in Fits program's --verify info
In the book, it was added that with the `--quiet' option, users can supress
the extra information that is printed with the `--verify' option. Also, in
the program, the extra information was made to be more clear (with every
extra line being printed above the final verify outputs (not before and
after it).