Corrected call to fits_write_chksum in Fits program
When I initially added this CFITSIO function, I had a typo in the full
function name, so it couldn't compile and I was forced to use the short
name (`ffpcks'). I just found that the problem wasn't in CFITSIO (not
properly defining the long funciton name), but in my own typing of it: so
the long option name is now used correctly.