Copying of a range of keywords from one HDU to another with Fits
With this commit, the Fits program can copy a range of keyword records from
one HDU to another (possibly in another file).

To do this, we also had to define a blank macro for the `long' and
`unsigned long' types. Finally, since this involved a change in the library
(a start of many to come in future commits), the library Soname was
incremented to 7 (from 6).