Match can merge all columns of one input with other input
Until now, if you wanted all the columns of one input to be merged with
those of another when matching catalogs, you would have to list them all on
the command-line (which is not convenient).

With this commit, it is possible to pass a special `_all' name to the Match
program's `--outcols' option. When Match sees this, it will print all the
columns of the respective input table.