Arithmetic's reusing dataset name and collapse memory usage fixed
When re-using a name for a different dataset, until now we would continue
parsing the full range of names, even after the old dataset with the
respective name had been freed! As a result, when there was more than one
named dataset, this would cause a segmentation fault. So a `break' has been
added to the `for' loop to stop the parsing after freeing the old dataset.

Also, while investigating this issue, I found another problem with multiple
uses of the `collapse-XXXX' operators in Arithmetic: if it was called more
than once (on different datasets, it would just continue decreasing the
dimensions of the reference dataset to 0 and negative, causing another
segmentation fault. So before decreasing the reference dataset's
dimensions, we now check with the dimension of the collapsed dataset.

This fixes bug #55439.