--checkconfig prints parsing of configuration files
With this option, its now possible to check/debug how options are parsed by
any Gnuastro program in the same order that they are parsed.

This idea for this option actually came up when I was checking how
`--minmapsize' is being parsed by the Fits program. I found that
`--minmapsize' is being ignored and that there is no check when its value
is zero (it is needed for allocating some datasets). To address the first,
its no longer in the list of hidden options for Fits and for the second, we
there is now a warning to inform the users of unexpected behavior.