Directory for writting memory-mapped files now checked
Until now, beyond its existance, there wasn't actually any checks on the
directory to host the memory-mapped files. Therefore, in cases like when it
was a failed symbolic link, or with no write permissions, the program that
needs to map memory would crash.

With this commit, we now check the writability of this directory also, and
if it exists and we don't have write permission, then another
`.gnuastro_mmap' directory will be created. If that doesn't exist either,
then the memory-mapped file will be in the current directory. In short, we
don't want a program to crash because of not being able to create the
memory mapped file.

This bug was reported by Raúl Infante Sainz.

This fixes bug #55478.