ConvertType: forced range also when using color maps
Until now, `--flminbyte' and `--flmaxbyte' were used to force the range of
conversion to color channels (even if the range is beyond the limits of the
dataset). With the introduction of color maps recently, it is also
necessary to force a range on non-byte datasets. It is thus necessary to
use a more generic name.

With this commit, it is now also possible to force the range in
single-input-channel, but color-mapped outputs.