NoiseChisel: good error message when quantile threshold not found
Until now, when no tiles could be used to find the quantile threshold,
NoiseChisel would crash with an error saying it can't allocate a zero-sized
array, not saying exactly why this was happening!

We had a very complete and descriptive error message *after* removing
outliers, but not before it. So unless the user ran NoiseChisel again with
`--checkqthresh' and inspected the extensions, they couldn't see if this
error occurred before the outlier rejection or after it.

With this commit, a function is defined to print the elaborate error
message while mentioning at which point the error happens. Its suggestions
are also increased if the error message happened after the outlier
rejection step (since there are two more parameters).

The particular crash of this bug was actually happening because
`gal_statistics_no_blank_sorted' was trying to copy a zero-sized array. So
with this commit, in such situations (when `size==0', it will just make a
blank array when `inplace==0' instead of attempting to copy the input.

Also, to implement this feature, the new `gal_blank_number' function was
defined (as a small wrapper around `gal_statistics_number') in the library.

This bug was reported by Raúl Infante Sainz.

This fixes bug #55491.