Arithmetic: new tofile- operator to write the top operand to a file
Until now, as the Arithmetic command got more and more complicated, there
it was relatively hard to check the progress and debug/understand
undexpected outputs.

With this commit, a new `tofile-AAA' operator is added to Arithmetic. It
can be placed anywhere in between the operands and operators to check the
top operand.