BuildProgram: --minmapsize no longer a hidden option
Recently a warning message was added to the library's option parsing
function to inform the users that `minmapsize' is not set or is set to zero
and discuss its implications.

But when an option is hidden, its value is not read or set. So BuildProgram
(which has set this option to hidden) would print this warning on every
run. Until now it was hidden to avoid the output of `--help' being too
extensive, but since we are using `gal_data_t' more and more, and also to
avoid this warning, it is now removed from the hidden list of options.