Arithmetic: --wcsfile to specify file of output's WCS
With this new option, its now possible to specify a different file
(possibly none of the inputs) to read the output's WCS from. It actually
came up while trying to fix bug #55544 (described below).

In the manual (version 0.8) we have mentioned that "output WCS information
will be taken from the first input image encountered". This would logically
mean the first dataset (with a WCS) that is found on the command-line.

However, for implementation purposes (especially when there are hundreds of
files), Arithmetic doesn't immediately read a file into memory! It will
keep its reference and load it into memory only when processing is to be
done on it (and immediately freed after it is no longer necessary).

This caused confusion when the "where" operator is used, for example in a
scenario like this:

$ astarithmetic image1.fits --hdu=1 image2.fits --hdu=1 \
                --output=masked.fits 1 eq nan where

In this scenario, while the `image1.fits' is the first one on the
command-line, `image2.fits' is the first one that is actually operated
on. So Arithmetic would load the WCS of `image2.fits' and not `image1.fits'
(as mentioned in the manual).

To fix this bug, the reading of the output's WCS is now separated from the
actual reading of the dataset arrays: it is now done when we add the files
to the stack (without read the actual data into memory).

This bug was reported by Raúl Infante-Sainz.

This fixes bug #55544.