Reporting the version of Bash if not suitable for Libtool
Very old versions of Bash fail with Libtool, but since many users expect
Bash to work with it anyway, it can be confusing when the configure script
just says that a proper shell couldn't be found. To help in debugging the
problem, if the version of Bash isn't good for libtool, it will be reported
(both as a new line during the tests and after it finishes).

This was suggested by David Valls-Gabaud.