NoiseChisel and Statistics: more generic --ignoreblankintiles
Until now, NoiseChisel and Statistics had a `--ignoreblankinsky' option
which would allow ignoring blank pixels in the tiled Sky and Sky standard
deviation outputs. But it was not possible to specify the same thing for
other tiled outputs. So the option name was generalized to
`--ignoreblankintiles' and is now used in the creation of any tiled output.