Arithmetic: New tofilefree operator and change in --dontdelete
Until now, as in all programs, when the `--dontdelete' option is called,
Arithmetic would simply complain and abort if the file exists. But with
this commit, only in Arithmetic, this option is interpretted as appending
the output dataset to the existing file's extensions.

Also, Arithmetic now has a new `tofilefree' operator that will free the
popped operand from the stack (as opposed to `tofile' which will put it
back ontop of the stack).

These features were added following a discussion with Leindert Boogaard.