NoiseChisel: New options to remove hard-coded configuration steps
Some steps were hardcoded in Gnuastro's code and thus couldn't be modified
by the user. With this commit, several of these configuration steps have
been opened up with the new `--dopening', `--dopeningngb'.

Also, the `--snthresh' option was added to allow the user to skip the step
to identify false pseudo-detections.