Interpolate library: reset flag to properly sort neighbors
Until now, in `gal_interpolate_close_neighbors', we weren't re-initializing
the flags of the interpolated neighbors prior to interpolation. As a
result, after finding the median value for the first time on each thread,
no other value was sorted any more (the median finding function would
assume it is already sorted because of the flag!).

This directly affected NoiseChisel in a very bad way: On the tiles where
outlier values fell right in the middle of the list of good neighbor
values, the outlier would be reported as the median (because the median is
the middle value!). In practice, this caused diamond like structures around
outlier tiles during NoiseChisel's threshold finding and Sky estimation
steps.

The fix to the problem was a single line before calling the median function
to reseting the sort and blank flags of the input dataset.

This bug came up during a discussion with Roberto Baena Gallé.

This fixes bug #55740.