Crop: Correct blank value for reading of non-basic FITS integer types
Until now, we were using the same allocated space pointer to write the
BLANK keywords and to use as blank values in the the dataset that CFITSIO
reads. For all basic types defined in the FITS standard this is
fine. However, for the non-simple ones (those that are defined with the
`BZERO' and `BSCALE' keywords), this would cause problems.

This bug was reported by Raúl Infante-Sainz.

This fixes bug #55763.