MakeProfiles: Warning when flat profiles will have a value of zero
It some times happens that we mistakenly don't write the input columns when
making a fast catalog properly. When there aren't enough columns,
MakeProfiles will complain and abort. But when an extra column is
mistakenly input, it can cause some unexpected behavior.

One such situation that has happened for myself several times happens when
making flat profiles with `--mforflatpix' or `--mcolisbrightness': when I
mistakenly add a 0 for the magnitude column. In this case, the flat profile
would get a value of zero, and it would thus be invisible in the final
image.

To avoid such unexpected behavior, a warning is now written at the start of
MakeProfiles when the scenario above happens. This will help the user be
aware of this situation and if they want to, they can disable it (along
with all input messages).