Library (interpolate): Ability to define the nearest-ngb metric
Until now, the nearest neighbor interpolation library would only use the
Manhattan/taxicab distance between two points. But this could cause small
`x'-like artifacts in the final result.

With this commit, its also possible to ask for a radial metric. As a
result, the programs that use this function (Arithmetic, NoiseChisel and
Statistics) also needed a new option and thus the `interpmetric' was added
as a tessellation option.

This solution was implemented after a discussion with Ignacio Trujillo.