CosmicCalculator: minmapsize option not hidden to avoid warning
Recently, a warning was added to the general option parsing library so when
`--minmapsize' has a value of zero or wasn't set, a notice would be
printed. This is very good to inform the user of possible performance
issues when this option is not set properly.

However, when an option is given a hidden attribute, it is no longer set,
so this warning is always printed. This was the case in CosmicCalculator
and this annoying warning would be printed all the time.

To solve this problem, with this commit, minmapsize is no longer a "hidden"
option.