Arithmetic: numbers printed with custom print format based on type
Until now, when printing a single number, Arithmetic would simply convert
the dataset to 64-bit floating point, then use `%g' to print it to standard
output. But that would make unreasonable results (for example a large
integer would be written in scientific notation which discards some of the
decimals).

This feature of Arithmetic pre-dates the `gal_type_to_string', so with this
commit we are using this function to print the value with the proper printf
formating in its native type.