Arithmetic: sqrt, log and log10 will always produce floating point
Until now, these three operators would save their results in the same type
as the input. But this is not a natural way to use these operators: they
don't map linearly and commonly loosing floating-point precision makes it
hard to deal with them.

With this commit, the Arithmetic libray's internal
`arithmetic_unary_function' will only do in place operations if the input
already has floating point types. If it doesn't, the output will be 32-bit
floating point.