Library (fits): gal_fits_name_save_as_string works with stdin
Until now, when this function was given a filename that is `NULL', it would
cause a segmentation fault in the program. But when the input is from the
standard input, the `filename' will be `NULL'. So it now returns `stdin' in
such cases.

This bug was reported by Elham Eftekhari.

This fixes bug #55845.