Table: New --sort and --range columns to select/modify rows
Until now, the Table program could only select certain input columns for
the output which was not convenient when you only need a subsection of them
and want to preserve the metadata (so a simple AWK call isn't an
option). With these two options, the output rows of Table can also be
selected and modified.

To enable this easily, the old `gal_qsort_index_single_i' and
`gal_qsort_index_single_d' functions have been generalized to accommodate
all numeric data types.