Table: checking --range columns from start of input table
In the previous commit, the new `--range' option was given to the Table
program. But in the step where we parse the input catalog's column names,
the first loop was over the input columns, not on the requested `--range'
columns. Therefore if the order of the two was different the search
failed.

With this commit, the first loop is over the columns given to `--range' and
all the input columns are searched for each `--range' column.

This bug was reported by Leindert Boogaard.