Match: printing newline after second output name in verbose mode
Until now, I had forgotten to print a new-line after the name of the second
output. On the command-line this would result in the terminal prompt being
on the same line and in a shell, the next outputs were printed on the same
line. Both caused confusion and is not expected

With this commit, a new-line is printed to fix this issue.