gal_eps_write: new dontoptimize argument to disable bit-optimization
Until now, bit-optimization (specific to EPS and PDF) was done on any
dataset with two values. However, in some situations (for example when the
user gives `--forcemin' or `--forcemax' to ConvertType), the user doesn't
want the contrast to be the full contrast between black and white.

With this commit, `gal_eps_write' has the new `dontoptimize' option to let
the caller determine if bit-optimization should be applied or not.