Library (tessellation): single tile with small remainder works
Until now, when the tile-size is very similar to the size of the dataset
along the given dimension, the tessellation library's produced tiles don't
fully cover the dataset. This affects NoiseChisel, Segment and Statistics.

With this commit, when setting the `last' tile-size in
`gal_tile_block_blank_flag', we first check if there is only one tile or
not. If there is indeed only one, the `last' will have the same size as the
`first'.

This fixes bug #56001.