Fits: Check for --datetosec, in case keyword couldn't be read
We recently added the `--datetosec' option to convert the date to the Unix
epoch time (in seconds). However, there was no check to see if the keyword
value could be read or not: causing confusing errors when the given keyword
didn't exist.

With this commit, a check has been added and Fits will crash when CFITSIO
couldn't parse the given keyword's value.