astscript-sort-by-night: better documentation
A few example customizations were given for this script. Also the file-name
format was more completely described in the description of `--link' and
`--prefix' is also explained better. Finally, I had mistakenly used `-h' as
the short option for `--hour'. This was wrong (because `-h' is the short
option for `--hdu'). This typo was also corrected.