Library (WCS): will not crash with singular WCS matrix
Until now, there was no check for a singular WCS matrix (where the
determinant is zero). But in one dataset, we noticed that this situation
has occurrred, so there were bad warnings and crashs when running
NoiseChisel.

With this commit, in such situations, the pixel scale (which caused those
errors) will not be calculated and the higher-level functions that used it
are also able to avoid the problem. Ofcourse a problem may arise in the
interpretation of the WCS and in that case, it should indeed crash. But in
writing a dataset, it shouldn't.

Also, while testing with the dataset (in the middle of the checks above), I
noticed that there are no checks to avoid a crash in Fits when called with
`--copykeys', but without any output filename or HDU.