astscript: better error message when on-off option is given value
Until now, we weren't actually exiting the configure script when a value
was given to an option that doesn't take one the the installed
scripts. With this commit, it exists in the function reporting the error
and the error message also prints the short option format name also for
better clarity.

I also noticed that the `-?*' series fitted many formats, so it was changed
to `-'?'*' which fixed the problem.