Full text of GNU Free Documentation License added
Until now, we were missing the full text of the GNU FDL in the version
controlled source. The Texinfo formatted version of the GNU FDL was
included in the manual after bootstrapping from Gnulib, but not in the
version controlled source.

With this commit, following the GNU recommendation (to add the license name
in capital letters, e.g., COPYING.LESSER), the full text of the GNU FDL is
added as `COPYING.FDL'.