astscript-sort-by-night: AWK command not using -v
Until now, this installed script's AWK command used the `-v' option to
define a variable on the command-line. However, the `-v' option is not
available on some AWK implementations.

With this commit, the calls to AWK don't use this option so it can be more
portable.

Also, the `tests/during-dev.sh' script has been modified to make it easy to
test the installed scripts.

This bug was reported by David Valls-Gabaud.