Book: redistributed installed script and astscript-sort-by-night sections
Until now, we had a separate chapter for installed scripts. This chapter
was intended to house the installed scripts. However, in the manual, we
have separate chapters for different types of operations. Therefore having
the scripts in a separate chapter (independent of the operation they do)
was confusing.

With this commit, the general introduction to installed scripts has been
moved to the "Common program behavior" chapter and the "Sort FITS files by
night" section has been moved to the data containers chapter.