NoiseChisel: cleaner check on neighbor-related options
Until now, the neighbor-related options like `--erodengb' were checked
individually to make sure their value is reasonable (either 4 or 8 in a 2D
image). But this was very dense checks, potentially making bugs.

With this commit, the new `ui_ngb_check' function has been added to do this
check for any option, after reading the dimensionality of the data.