NoiseChisel: S/N table writing won't crash when there are no rows
Until now, when there were no rows in the S/N table (with `--checksn')
NoiseChisel would crash with a statement saying that this is a
bug. But this isn't a bug! Because tables with zero rows can be built.

With this commit, the check that crashed NoiseChisel has been removed to
address this problem.