Statistics: check if single-valued measurement options have a value
Until now, there was no check in statistics to see if single-valued options
that need a value (like `--quantile') have one. This can cause
unpredictable behavior and is not good. Statistics should complain and
abort in such cases.

With this commit a check has been added and Statistics will abort when it
confronts such a case.

This fixes bug #56246.