Statistics: Sigma-clipping results as single-valued measurements
Until now, the only way to get Sigma-clipped results from Statistics was to
use the `--sigmaclip' option. It prints a lot of information along with the
final result. This is useful for visual inspection, but makes automatic use
of the output a little hard.

With this commit, Statistics has four new options to print the
sigma-clipped results as single numbers: `--sigclip-number',
`--sigclip-median', `--sigclip-mean', `--sigclip-std'. A fully working
demonstration was also added in the book (under `--sigclip-median') on how
to use it to select certain rows of a table.

While doing this a few other issues were addressed:

 - I noticed that Table's `--range' option isn't removing blank pixels.

 - I also noticed that the `gal_statistics_no_blank_sorted' wasn't taking
   `inplace' into account when the array is already sorted.

 - Since this is the first commit, working in the library after the
   previous release, the library soname version was incremented to 8.

This was suggested by Raul Infante-Sainz.