MakeCatalog: better description for the semi-major/minor options
Until now, when describing the two `--semimajor' and `--semiminor' options,
we were just saying "Flux weighted semi-XXXX axis". But this is not exactly
clear on how it is derived and could cause confusions.

With this commit, we now say "RMS along XXXX axis (in pixels)" to be clear
and avoid confusion.

This was suggested by Lee Kelvin.