ConvertType: non-color output formats don't scale to uint8
Just before release 0.8, we added colormaps to ConverType. They are very
useful when dealing with output formats that accept color. However, when
they don't (for example converting a plain text array to FITS), ConvertType
still scales the range to unsigned 8-bit integer.

With this commit, this scaling is now taken only in the parts of formats
that accept color. When they don't, the dataset is written to the output as
it is.

Also, while doing this, a problem with reading plain text files as an array
was also fixed: instead of requiring a format for every "column", when the
array is an image, a single format is used all the time.

This fixes bug #56256 and bug #56257.