Arithmetic: coadding operators return float32 and uint32 (for counting)
Until now, the returned dataset from the coadding operators had the same
type as the input datasets. This would cause a lot of information (for
example when the inputs are integers and we want to coadd by mean, which
needs float), or have too much information (for example when the inputs are
float and you want to coadd by number, which needs integer).

To solve this problem, with this commit, the output type is determined by
the operator, not the input data type.

This was suggested by Raul Infante-Sainz.