gal_statistics_outlier_flat_cfp: new implementation and API
Until now, the arguments to this function were actually percentages, based
on the range of the dataset's values. But when a dataset has huge outliers,
such percentages are worthless and can be very different with similar
datasets but without the outliers. The threshold for flatness was also an
absolute value, making it hard to generalize.

With this commit, this function now uses the previous elements as a
reference to define flatness. Also the criteria on width is now in units of
numbers (of elements), not a percentage of the the range of the input
dataset.