CosmicCalculator: redshift zero request is converted to 1e-14
Until now, CosmicCalculator would fail when given a value of zero with this
GSL error: "gsl: qng.c:189: ERROR: failed to reach tolerance with
highest-order rule". So its numerical integration would fail to converge.

To fix the problem, with this commit, when the user requests results for
z=0, we internally convert it to `1e-14' (close to, but not exactly, the
64-bit floating point precision limit).

This bug was reported by Leindert Boogaard.

This fixes bug #56299.