CosmicCalculator: find redshift based on observe wavelength
Until now the only way to specify a redshift in CosmicCalculator was to
explicitly give it with the `-z' option. But in some situations, you know
the observed wavelength of an emission line, and its rest-frame wavelength
(which can also be used to estimate the redshift).

With this commit CosmicCalculator has the new `--obsline' option where you
can feed this information about observed and rest-frame wavelengths and let
CosmicCalculator estimate the redshift from them and continue as
before. Since we aren't explicity giving the redshift, another
`--usedredshift' option has been added to CosmicCalculator to print the
used redshift with any other requested calculated value.

In order to do this, it was necessary to define a new library in Gnuastro
to keep the wavelengths and give standard names to spectral lines. That
library is now saved in `lib/speclines.c' and `lib/gnuastro/speclines.h'.