Book: listed all lines in CosmicCalculator's --obsline description
Many new lines were added since the first description for `--obsline', but
I had forgot to add them to the description of the `--obsline' option. With
this commit, they have been added.

Also, I noticed that there were no index entries for the line macros in the
library section of the book. So index entries were also added.