Match: --stdintimeout option is not ignored
Match is designed to also work based on piped input from `stdin'. However,
until now, the `--stdintimeout' option was set as a "hidden" option, so its
value from the configuration files would be ignored and Match would give it
a value of zero. So in many cases, it couldn't read input from `stdin'.

With this commit, it is no longer set as a hidden option.