txt.h: optionally parse meta-data without modifying the line
Until now, when there was a meta-data line (starting with `# Column N:'),
the `txt_info_from_comment' function would modify the contents of the line
(in partciular, it would put string-NULL characters in it for the different
components that it needs). As a result, a program like Match that would
need to parse these meta-data lines again wouldn't find any.

With this commit, an `inplace' argument has been added to this option, so
when it is zero, a separately allocated line is parsed and modified, not
the original one.

This fixes bug #56324.