Book: building and installing shared CFITSIO library
Until now, the build instructions for CFITSIO as a dependency of Gnuastro
only built it as a static library. As a result, if the user's computer had
other shared library installations of CFITSIO (with different versions),
that would cause an error like below (after the main body of the commit
message).

After many tests with the help of Elham Saremi, we were able to track down
this problem and fix it. The fix was to build and install a shared CFITSIO
library also (not just the default static library). As a result with this
commit, CFITSIO's installation instructions include `make shared'.

"""""
ERROR: Mismatch in the CFITSIO_SONAME value in the fitsio.h include file
that was used to build the CFITSIO library, and the value in the include
file that was used when compiling the application program:
   Version used to build the CFITSIO library   = 2
   Version included by the application program = 7

Fix this by recompiling and then relinking this application program
with the CFITSIO library.
""""