Book: Corrections and edits to the Sufi tutorial
Hamed Altafi pointed out a few typos in this tutorial (mainly things we
forgot to update after the program's new features), and had some advice
(e.g., not introducing Sufi at the start of the tutorial, and advising
caution when copying the catalog from the PDF book).

With this commit, all the raised points have been implemented.