Installation: More robust checks of libraries with AC_LIB_LINKFLAGS
Until now, we only check for the link-ability of a library. But it was
recently pointed out by Bruno Haible that we can use Gnulib's
`AC_LIB_LINKFLAGS' to add `rpath', while doing better tests on the properly
found libraries.

With this commit, the checking of libraries now uses this Gnulib autoconf
check which actually includes compiling and linking a small test
program. Also, the test for the Steffen library in GSL now uses the much
more robust `AC_LINK_IFELSE' Autoconf test which actually tries compiling
and linking a small test program also. This will allow avoiding many
frustrating bugs when multiple versions of GSL are available in the system.

This test was done with the help of Elham Saremi and Raul Infante Sainz.