Book: dealing with time zones in sort-by-night script
With this commit, documentation about how to deal with difference in
time zones when using `sort-by-night' script has been improved.
Basically, it is explained how to modify the definition of "nights" when
clasifying and sorgint images taken from different nights with the help
of the argument `--hour'.

Since this commit has been done in a Mac OS system laptop, I also
include `.DS_Store' files to be ignored in `.gitignore'. This files are
created automatically by `Finder' application and they are irrelevant
for the git history. From Wikipedia: In the Apple macOS operating
system, .DS_Store is a file that stores custom attributes of its
containing folder, such as the position of icons or the choice of a
background image.