Book: minor edits in newly added parts
In order to make the text of the book more smooth, the text added in the
previous commit was edited to fit into the same style as the rest of the
book. In particular, "argument" was corrected to "option", along with a few
minor style issues. Also, some index entries were added. A similar edit was
done in the `--help' introduction.

Finally, when building Gnuastro, I noticed that we hadn't included
`astscript*' files as files to be ignored. So this was also done with this
commit.