Build: three compiler warnings have been addressed
Raul Infante-Sainz reported that Gnuastro's build on macOS produced two
warnings in Convolve and one in the libraries. With this commit, they are
both corrected.

In the first case, we were not writing the output string into `source' when
there was a file as input. In the second, we were checking a condition that
is always true: the address of an allocated space. The third was due to an
initialization (irrelevant to the program!).